Move isvc tests into predictor_test.go; move TLS tests to end
